Drop Item Bug
Fixed a bug where player could not drop an item in the enemy's detection radius. Still cannot drop an item on top of an enemy (i.e. if the enemy is directly behind the player), but that is fine--a feature, not a bug. It could be corrected by giving the enemy's hitbox its own layer and having the PlayerItemHitbox ignore that layer.